Specialise the property -- the number of lines must be odd.
This implies greater than zero, so I changed the test rather than adding a new one.

But the test still passes: my current implementation returns a single line, and 1 is an odd number.  Not a great choice of property.

I checked that the test failed if the property was that the number of lines should be even, so I'm happy the test works. But it does not help move the implementation forward.

I'll keep the new test anyway, since it works and does constrain the implementation more tightly.