A *failing* test to drive development forward...
Instead of specifying that the number of lines is odd, define how the number of lines relates to the letter parameter.

This introduces a new domain concept 'ord(c:Char):Int'.  This may come in handy as we grow the implementation, but for now I'll keep it in the test code.

Note the quality of the failure messages that the ScalaTest assert macro generates!